Refactor get_task_stream
